requirements.txt
flask
openai
transformers
torch
faiss-cpu
pandas
scikit-learn
python-docx
PyPDF2
